updated for prod

+++ Models/ObjectModel.cs
@@ -7,11 +7,11 @@ namespace HalogenPreTestAPI.Models;
 
 public class ObjectModel
 {
-    public List<Int32>? DivBy3 { get; set; }
-    public List<Int32>? DivBy5 { get; set; }
-    public List<Int32>? DivBy7 { get; set; }
-    public List<Int32>? EvenNums { get; set; }
-    public List<Int32>? OddNums { get; set; }
-    public int Mode { get; set; }
+    public List<Double>? DivBy3 { get; set; }
+    public List<Double>? DivBy5 { get; set; }
+    public List<Double>? DivBy7 { get; set; }
+    public List<Double>? EvenNums { get; set; }
+    public List<Double>? OddNums { get; set; }
+    public double Mode { get; set; }
     public double Median { get; set; }
 }
